+++ docs/conf.py
@@ -27,7 +27,7 @@ author = 'Brian Cappello'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = 'v0.8.0'
+release = 'v0.8.1'
 
 
 # -- General configuration ---------------------------------------------------